+++ source/monitor/NrpTimer.h
@@ -10,13 +10,13 @@ namespace gui
 class CNrpTimer : public IGUIAnimator
 {
 public:
-	CNrpTimer( IGUIEnvironment* environment, IGUIElement* node, int time, core::stringc action );
+	CNrpTimer( IGUIEnvironment* environment, IGUIElement* node, size_t time, core::stringc action );
 	~CNrpTimer( void );
 
 	virtual void draw();
 private:
-	int _time;
-	int _startTime;
+	size_t _time;
+	size_t _startTime;
 	bool _active;
 	core::stringc _action;
 };
